Move review scope
review belongs to the change endpoint moving
to be aligned with Gerrit REST API Docs.
